Rename file to DalAstHelper to be more accurate.
